+++ cmd/autothrottle/throttles.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
-	"log"
 
 	"github.com/DataDog/kafka-kit/kafkametrics"
 	"github.com/DataDog/kafka-kit/kafkazk"
@@ -135,144 +133,6 @@ func (r replicationCapacityByBroker) setAllRatesWithDefault(ids []int, rate floa
 		r.storeLeaderCapacity(id, rate)
 		r.storeFollowerCapacity(id, rate)
 	}
-}
-
-// updateReplicationThrottle takes a ReplicationThrottleConfigs that holds
-// topics being replicated, any ZooKeeper/other clients, throttle override
-// params, and other required metadata. Metrics for brokers participating in
-// any ongoing replication are fetched to determine replication headroom.
-// The replication throttle is then adjusted accordingly. If a non-empty
-// override is provided, that static value is used instead of a dynamically
-// determined value. Additionally, broker-specific overrides may be specified,
-// which take precedence over the global override.
-func updateReplicationThrottle(params *ReplicationThrottleConfigs) error {
-	// Creates lists from maps.
-	srcBrokers, dstBrokers, allBrokers := params.reassigningBrokers.lists()
-
-	log.Printf("Source brokers participating in replication: %v\n", srcBrokers)
-	log.Printf("Destination brokers participating in replication: %v\n", dstBrokers)
-
-	// Determine throttle rates.
-
-	// Use the throttle override if set. Otherwise, make a calculation
-	// using broker metrics and configured capacity values.
-	var capacities = make(replicationCapacityByBroker)
-	var brokerMetrics kafkametrics.BrokerMetrics
-	var rateOverride bool
-	var inFailureMode bool
-	var metricErrs []error
-
-	if params.overrideRate != 0 {
-		log.Printf("A global throttle override is set: %dMB/s\n", params.overrideRate)
-		rateOverride = true
-
-		capacities.setAllRatesWithDefault(allBrokers, float64(params.overrideRate))
-	}
-
-	if !rateOverride {
-		// Get broker metrics.
-		brokerMetrics, metricErrs = params.km.GetMetrics()
-		// Even if errors are returned, we can still proceed as long as we have
-		// complete metrics data for all target brokers. If we have broker
-		// metrics for all target brokers, we can ignore any errors.
-		if metricErrs != nil {
-			if brokerMetrics == nil || incompleteBrokerMetrics(allBrokers, brokerMetrics) {
-				inFailureMode = true
-			}
-		}
-	}
-
-	// If we cannot proceed normally due to missing/partial metrics data,
-	// check what failure iteration we're in. If we're above the threshold,
-	// revert to the minimum rate, otherwise retain the previous rate.
-	if inFailureMode {
-		log.Printf("Errors fetching metrics: %s\n", metricErrs)
-
-		// Increment and check our failure count against the configured threshold.
-		over := params.Failure()
-
-		// If we're not over the threshold, return and just retain previous throttles.
-		if !over {
-			log.Printf("Metrics fetch failure count %d doesn't exeed threshold %d, retaining previous throttle\n",
-				params.failures, params.failureThreshold)
-			return nil
-		}
-
-		// We're over the threshold; failback to the configured minimum.
-		log.Printf("Metrics fetch failure count %d exceeds threshold %d, reverting to min-rate %.2fMB/s\n",
-			params.failures, params.failureThreshold, params.limits["minimum"])
-
-		// Set the failback rate.
-		capacities.setAllRatesWithDefault(allBrokers, params.limits["minimum"])
-	}
-
-	// Reset the failure counter. We may have incremented in past iterations,
-	// but if we're here now, we can reset the count.
-	if !inFailureMode {
-		params.ResetFailures()
-	}
-
-	// If there's no override set and we're not in a failure mode, apply
-	// the calculated throttles.
-	if !rateOverride && !inFailureMode {
-		var err error
-		capacities, err = brokerReplicationCapacities(params, params.reassigningBrokers, brokerMetrics)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Merge in broker-specific overrides if they're part of the reassignment.
-	for id := range params.reassigningBrokers.all {
-		if override, exists := params.brokerOverrides[id]; exists {
-			rate := override.Config.Rate
-			// A rate of 0 means we intend to remove this throttle override. Skip.
-			if rate == 0 {
-				continue
-			}
-			log.Printf("A broker throttle override is set for %d: %dMB/s\n", id, rate)
-			// Any brokers with throttle overrides that are being issued as part
-			// of a reassignemnt should be marked as such.
-			override.ReassignmentParticipant = true
-			params.brokerOverrides[id] = override
-			// Store the rate for both inbound and outbound traffic.
-			capacities.storeLeaderAndFollerCapacity(id, float64(rate))
-		}
-	}
-
-	//Set broker throttle configs.
-	events, errs := applyBrokerThrottles(params.reassigningBrokers.all, capacities, params.previouslySetThrottles, params.limits, params.zk)
-	for _, e := range errs {
-		log.Println(e)
-	}
-
-	//Set topic throttle configs.
-	if !params.skipTopicUpdates {
-		_, errs = applyTopicThrottles(params.reassigningBrokers.throttledReplicas, params.zk)
-		for _, e := range errs {
-			log.Println(e)
-		}
-	}
-
-	// Append broker throttle info to event.
-	var b bytes.Buffer
-	if len(events) > 0 {
-		b.WriteString("Replication throttles changes for brokers [ID, role, rate]: ")
-
-		for e := range events {
-			b.WriteString(fmt.Sprintf("[%d, %s, %.2f], ", e.id, e.role, e.rate))
-		}
-
-		b.WriteString("\n")
-	}
-
-	// Append topic stats to event.
-	b.WriteString(fmt.Sprintf("Topics currently undergoing replication: %v", params.topics))
-
-	// Ship it.
-	params.events.Write("Broker replication throttle set", b.String())
-
-	return nil
 }
 
 // brokerReplicationCapacities traverses the list of all brokers participating
